requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
docopt==0.4.0
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
mailchimp==2.0.10
mailchimp-marketing==3.0.75
MarkupSafe==2.1.1
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
